Move TLLog API
Next to Orga and TL, there is an API for marking logs.
The API is only used by the Orga team, but warrants its own section, since it's self-contained
