+++ bp/views.py
@@ -193,55 +193,3 @@ class StudentImportView(PermissionRequiredMixin, FormView):
                 print(f"Teilnehmer*in {row['Vollständiger Name']} existiert bereits")
         messages.add_message(self.request, messages.SUCCESS, f"{import_count} Teilnehmende erfolgreich importiert")
         return super().form_valid(form)
-
-
-class APILogMarkReadView(LoginRequiredMixin, DetailView):
-    http_method_names = ['post']
-    model = TLLog
-
-    def post(self, request, *args, **kwargs):
-        if request.user.is_superuser:
-            log = self.get_object()
-            log.read = True
-            log.save()
-            return HttpResponse("")
-        return HttpResponseForbidden("")
-
-
-class APILogMarkHandledView(LoginRequiredMixin, DetailView):
-    http_method_names = ['post']
-    model = TLLog
-
-    def post(self, request, *args, **kwargs):
-        if request.user.is_superuser:
-            log = self.get_object()
-            log.handled = True
-            log.save()
-            return HttpResponse("")
-        return HttpResponseForbidden("")
-
-
-class APILogMarkGoodView(LoginRequiredMixin, DetailView):
-    http_method_names = ['post']
-    model = TLLog
-
-    def post(self, request, *args, **kwargs):
-        if request.user.is_superuser:
-            log = self.get_object()
-            log.good_log = True
-            log.save()
-            return HttpResponse("")
-        return HttpResponseForbidden("")
-
-
-class APILogMarkBadView(LoginRequiredMixin, DetailView):
-    http_method_names = ['post']
-    model = TLLog
-
-    def post(self, request, *args, **kwargs):
-        if request.user.is_superuser:
-            log = self.get_object()
-            log.good_log = False
-            log.save()
-            return HttpResponse("")
-        return HttpResponseForbidden("")
